User interface Design
User interface (UI) design is about building interfaces with a focus on styling and interactivity. When we talk about UI design, we talk about the stylistic choices a designer makes when creating a product, e.g., an image, button, menu bar, or footer. All of these elements will affect the user’s interaction, and so, your goal is to create an interface 
the user finds easy to use and aesthetically pleasing.